--- steven_working_dir/index.hpp
@@ -1,47 +1,0 @@
-#ifndef INDEX_H
-#define INDEX_H
-
-#include <vector>
-#include <optional>
-// Assuming that we need some form of data structure for the indices, e.g., B-Trees
-// For simplicity, using std::map as a placeholder. Real implementation may vary.
-#include <map>
-
-// Forward declaration to resolve circular dependency
-class Table;
-
-class Index {
-private:
-    std::vector<std::optional<std::map<int, int>>> indices; // Placeholder for actual index structures
-    Table* table; // Pointer to associated table
-
-public:
-    // Constructor
-    Index(Table* table) : table(table) {
-        // Initialize indices based on the number of columns in the table
-        // Assuming Table class has a method `getNumColumns` to get the number of columns
-    // indices.resize(table->getNumColumns());
-    }
-
-    // Locate all records with the given value in a column
-    void locate(int column, int value) {
-        // Implementation goes here
-    }
-
-    // Returns the RIDs of all records with values in a column within a range
-    void locate_range(int begin, int end, int column) {
-        // Implementation goes here
-    }
-
-    // Optional: Create index on a specific column
-    void create_index(int column_number) {
-        // Implementation goes here
-    }
-
-    // Optional: Drop index of a specific column
-    void drop_index(int column_number) {
-        // Implementation goes here
-    }
-};
-
-#endif // INDEX_H
